+++ packages/general/src/crypto/index.ts
@@ -4,6 +4,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+export * from "./aes/Ccm.js";
 export * from "./Crypto.js";
 export * from "./CryptoConstants.js";
 export * from "./CryptoError.js";