+++ CHANGELOG.md
@@ -20,6 +20,9 @@ The main work (all changes without a GitHub username in brackets in the below li
 -   @matter/nodejs
     - Enhancement: Use "stat" to determine storage file existence instead of reading all content
 
+-   @matter/react-native
+    - Fix: (Luxni) Update UDP, BLE and Crypto usage to work with React Native
+
 ## 0.15.3 (2025-08-01)
 
 -   @matter/protocol